separate trait and amanger
